+++ src/lib/utils.ts
@@ -38,8 +38,18 @@ export const error = (): never => {
 export function getCssVariable(container: HTMLDivElement, property: string): string {
 	// Extract the value of the CSS variable from the container's computed style
 	const value = getComputedStyle(container).getPropertyValue(property).trim();
-
 	return value;
+}
+
+export function hex8ToRgba(hex: string): string {
+	const value = hex.replace('#', '');
+
+	const r = parseInt(value.slice(0, 2), 16);
+	const g = parseInt(value.slice(2, 4), 16);
+	const b = parseInt(value.slice(4, 6), 16);
+	const a = Math.round((parseInt(value.slice(6, 8), 16) / 255) * 100) / 100;
+
+	return `rgba(${r}, ${g}, ${b}, ${a})`;
 }
 
 export function mulberry32(seed: number) {
